[deploy] namespace and rename
This puts deploy into the torch::deploy namespace. It also renames some
objects to better match their behavior:

PythonObject -> Obj, in the future it will refer to either a python object or a handle to a script obj, so rename it torch::deploy::Obj to be generic
MovableObject -> ReplicatedObj, to prevent confusion with "std::move" which is unrelated, and to note that we are replicating this object across interpreters.

ghstack-source-id: aa56f6a4e99e0a8ba8dda035c416e19d61a5b22f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53670