Marginally improve pytest collection for top-level test files
I'm trying to make `pytest test/*.py` work--right now, it fails during
test collection.  This removes a few of the easier to fix pytest
collection problems one way or another.  I have two remaining problems
which is that the default dtype is trashed on entry to test_torch.py and
test_cuda.py, I'll try to fix those in a follow up.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 58b65d1fc573cfc0784af4d073472aa810cb3d5f
Pull Request resolved: https://github.com/pytorch/pytorch/pull/53617